Renamed search page. Only display search page to admins (for now). Hide action button on user list if it's the currently logged in user.
