+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "prime-solo-project",
-  "version": "0.1.0",
+  "version": "1.0.1",
   "private": true,
   "proxy": "http://localhost:5000",
   "dependencies": {
