+++ src/components/SearchPage/SearchPage.css
@@ -50,14 +50,7 @@
     width: 100px;
 }
 
-.reset-location{
-    margin-left: 10px;
-    padding-bottom: 30px;
-}
-
 .leader-board-button-div{
-    padding-left: 10px;
-    padding-top: 10px;
     border-top: solid #0c495a;
     border-top-width: medium;
 }
